tts-service/requirements.txt
fastapi
uvicorn
kafka-python
boto3
gtts
pyyaml
